# specs/api.yaml
openapi: 3.0.0
info:
  title: Luna Creators Portal API Docs
  description: 'basePath is /api'
  version: 1.0.0
servers:
  - url: 'https://creators-portals-api-staging-8lv8j.ondigitalocean.app/api'
...
paths:
...
  /organizations:
    post:
      tags:
        - organization
      security:
        - ApiKeyAuth: [ ]
      operationId: addOrganization
      summary: Add Organization
      requestBody:
        content:
          multipart/form-data: # Media type
            encoding:
              image:
                contentType: image/png, image/jpeg
              banner:
                contentType: image/png, image/jpeg
            schema: # Request payload
              type: object
              required:
                - name
                - type
              properties:
                name:
                  type: string
                  minLength: 3
                  maxLength: 30
                type:
                  $ref: '#/components/schemas/OrganizationType'
                website:
                  type: string
                  format: uri
                twitter:
                  type: string
                  format: uri
                discord:
                  type: string
                  format: uri
                email:
                  type: string
                  format: email
                admin_name:
                  type: string
                admin_wallet_address:
                  type: string
                  pattern: (\b0x[a-fA-F0-9]{40}\b)
                image:
                  type: string
                  format: binary
                banner:
                  type: string
                  format: binary

      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateOrganizationResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
              example:
                error: true
                code: 403
                message: An admin auth key is required
    get:
      tags:
        - organization
      security:
        - ApiKeyAuth: [ ]
      operationId: getOrganizationsInfo
      summary: Get all organizations info
      parameters:
        - name: name
          description: part of the organization name
          in: query
          schema:
            type: string
            minLength: 2
          required: false
        - name: type
          description: organization type
          in: query
          schema:
            $ref: '#/components/schemas/OrganizationType'
          required: false
        - name: email
          description: part of the admin email
          in: query
          schema:
            type: string
            minLength: 3
          required: false
        - name: page
          in: query
          schema:
            type: number
            format: int32
            default: 1
          required: false
        - name: size
          in: query
          schema:
            type: number
            format: int32
            default: 30
          required: false
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAllOrganizationsInfoResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
              example:
                error: true
                code: 403
                message: An admin auth key is required
  /organizations/{organization_id}:
    parameters:
      - name: organization_id
        in: path
        schema:
          type: string
          format: uuid
        required: true
    put:
      tags:
        - organization
      security:
        - ApiKeyAuth: [ ]
      operationId: updateOrganization
      summary: Update organization info
      requestBody:
        content:
          multipart/form-data: # Media type
            encoding:
              image:
                contentType: image/png, image/jpeg
              banner:
                contentType: image/png, image/jpeg
            schema: # Request payload
              $ref: '#/components/schemas/UpdateOrganizationRequest'
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateOrganizationResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
              example:
                error: true
                code: 403
                message: An admin auth key is required
    get:
      tags:
        - organization
      security:
        - ApiKeyAuth: [ ]
      operationId: getOrganization
      summary: Get organization
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetOrganizationResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
              example:
                error: true
                code: 403
                message: An admin auth key is required
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-auth-token
    PluginAuth:
      type: http
      scheme: basic
  schemas:
    GenericResponse:
      properties:
        error:
          type: boolean
          example: false
        code:
          type: number
          example: 200
        message:
          type: string
          example: Successful
    Pagination:
      type: object
      properties:
        page:
          description: First page is page 1, default size if not specified is 50.
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        last_page:
          description: Last possible page
          type: integer
          format: int32
        total_count:
          type: integer
          format: int32
      example:
        page: 1
        size: 50
        last_page: 5
        total_count: 241
...
    OrganizationType:
      type: string
      enum:
        - BRAND
        - COMMUNITY
...
    UpdateOrganizationRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 30
        type:
          $ref: '#/components/schemas/OrganizationType'
        website:
          type: string
          format: uri
        twitter:
          type: string
          format: uri
        discord:
          type: string
          format: uri
        email:
          type: string
          format: email
        public_address:
          type: string
          pattern: (\b0x[a-fA-F0-9]{40}\b)
        image:
          type: string
          format: binary
        banner:
          type: string
          format: binary
    OnboardingType:
      type: string
      enum:
        - INVITED
        - ADMIN_CREATED
    OrganizationInfo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nonce:
          type: number
        onboarding_type:
          $ref: '#/components/schemas/OnboardingType'
        name:
          type: string
          minLength: 3
          maxLength: 30
        website:
          type: string
          format: uri
        twitter:
          type: string
          format: uri
        discord:
          type: string
          format: uri
        facebook:
          type: string
          format: uri
        instagram:
          type: string
          format: uri
        email:
          type: string
          format: email
        public_address:
          type: string
          pattern: (\b0x[a-fA-F0-9]{40}\b)
        created_at:
          type: string
          format: 'date-time'
        updated_at:
          type: string
          format: 'date-time'
        image:
          type: string
          format: uri
        banner:
          type: string
          format: uri
    UpdateOrganizationResponse:
      allOf:
        - $ref: '#/components/schemas/GenericResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/OrganizationInfo'
    GetAllOrganizationsInfoResponse:
      allOf:
        - $ref: '#/components/schemas/GenericResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                pagination:
                  $ref: '#/components/schemas/Pagination'
                items:
                  $ref: '#/components/schemas/OrganizationInfo'
    GetOrganizationResponse:
      allOf:
        - $ref: '#/components/schemas/GenericResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/OrganizationInfo'
...